pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-06"
version = "0.1.0"
description = ""
authors = ["filimon4uck <filimon4uck.yura@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.38"
faker = "^36.1.1"
psycopg2 = "^2.9.10"
alembic = "^1.15.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
